refactor(notify): 移除 Lombok @Data，改用显式 Getter/Setter
- 全面替换 NotifyProperties 及其内嵌配置类中的 @Data 注解，改为手写 getter 和 setter 方法
- 统一配置类风格，提升代码可读性和维护性
- 新增邮件渠道安全配置、安全追踪、DKIM 签名等功能支持
- 修正邮件 SSL 配置中 trustStore 字段命名错误
- 调整 DedupConfig 中窗口时间属性名称，确保命名一致性
- 优化 NotifyDedupService 中去重时间窗口类型，避免隐式类型转换问题
- 精简 MessageAutoConfiguration 中 CachedTemplateEngine 的 bean 创建方法签名
- 优化 ExcelFacade 新增支持读取全部 Sheet 的无类型原始数据功能
- 修正 AuthContextUtils 中 TenantContextHolder 设置逻辑，使用 Builder 模式实例化
- 完善 BatchServiceImpl 中批量任务处理及异常捕获逻辑，加强容错性

+++ ydsz-common/ydsz-common-notify/src/main/java/com/njydsz/common/notify/dedup/NotifyDedupService.java
@@ -63,7 +63,7 @@ public class NotifyDedupService {
       return false;
     }
     String fingerprint = computeFingerprint(receiver, title, content);
-    int windowSeconds = properties.getDedup().getWindowSeconds();
+    long windowSeconds = properties.getDedup().getWindowSeconds();
     String redisKey = properties.getDedup().getRedisKeyPrefix() + fingerprint;
 
     if (redisStringOps != null) {
